SensorX2car/camera2car/auto_calib/requirements.txt
cython
torch>=1.5.0
torchvision>=0.6.0
scipy
scikit-image
tqdm
opencv-python
matplotlib
yacs
tensorboard
pylsd-nova